Add and fix out-of-bounds for delegated db vector
Also add multiple tests.

It turns out that we don't have to set the values in cache to `None` as
the `length` field handles all bounds checks for get, set, and pop. So
we could actually change the type of the cache from HashMap<IndexType,
Option<T>> to HashMap<IndexType, T>.