Remove `mut` from some delegated DB vec trait functions
Although an implementation uses a leveldb database under the hood and
all *reading* from the DB takes a mutable DB object, the reading
functions in the trait do not have to be mutable since the database is
wrapped in an Rc<RefCell<.>> smart pointer where wrapped struct can be
modified without the needing to be `mut`.

This means that we get an intuitive interface: The methods that would
mutate the length or values in the persisted vector take a mutable
reference to `self` whereas those that only read from it take `&self`.
The reason that the writing function *need* to take &mut self even
though the database is wrapped in Rc is that those functions need to
write to `cache` which is a regular `HashMap` that lives in memory.