feat: use rs-leveldb instead of rusty-leveldb
Addresses #163

rs-leveldb is a wrapper for C++ leveldb.  It is a fork of the standard
leveldb crate and seems nicer because it supports using strings and
standard integer types as keys directly.

With these changes the code is equally as functional and all tests pass.
Application locks remain in place, but are redundant to the locking that
occurs within C++ leveldb.  Our present Mutex locks harm concurrency, so
it is planned that a future commit will remove these.

Completes tasks:
[*] switch dep from rusty-leveldb to rs-leveldb and make code/tests compile.
[*] implement auto-destruction of database directories for tests
[*] get all tests passing without any test database directories persisting after test run.
[*] document deps to build with c++ impl.