test: attempt to make tests more reliable.
These two tests are non-deterministic:
 + non_atomic_set_and_get
 + non_atomic_set_and_get_wrapped_atomic_rw

They setup two threads with a race condition and panic when an
inconsistent state is detected.  If they don't panic, the test fails.

Sometimes the test has been failing on github CI.

The changes here are to increase test iterations (1000 --> 10000)
and introduce a 1 ms sleep in the reader thread.  Hopefully this
makes the tests pass at near 100% on CI.

fingers crossed!