Add Merkle tree tests for digest length 0 and 1
The test for the empty digest list currently fails. It can be fixed by
either adding a #should_panic attribute, or by changing the behavior of
`from_digests` to return an error on empty input, or by allowing empty
inputs in the Merkle tree construction. The latter option might be
problematic, as the height would have to be undefined, or -1.