Fix potential db-key collission for singleton and dbtvec length
By adding a deterministically derived value from the schema, the number
of tables added to the schema, and using this as a key for the underlying
leveldb database, we can ensure that no key collissions happen, as all
tables have a unique prefix that *all* their keys start with.

This closes #177.