Remove Tip5's custom-implementation of `sample_scalars`
The tasm-lib implementation of `sample_scalars` agrees with the
`sample_scalars` on the trait `AlgebrabraicHasher`, not the one defined
for `Tip5`. But Triton-VM seems to have relied on this `sample_scalars`
implementation that this commit deletes. The reason that this was able
to fly under the radar in e.g. `tasm-lib`'s `fri_verify` is that the Fri
verifier only ever sampled *one* x-field element at a time. Once you
start sampling more than three x-field elements, the two versions
(tasm-lib on one side vs. Triton-VM's and twenty-first's old one that's
deleted in this commit) would disagree.

A test I was writing in `tasm-lang` caught this discrepancy.

Co-authored-by: Ferdinand Sauer <ferdinand@neptune.cash>