Add benchmark for unsigned multiplying unsigned integers
This benchmark reveals that, at least on my Core i9-11950H processor,
the multiplications (u32,u32)->u64 and (u64,u64)->u128 take the *same*
amount of time. This result might help inform choices of prime fields
and might help estimate multiplication times for BFieldElements and
XFieldElements, so I'll add it to the list of benchmarks.