test: add tests for `BFieldCodec` derive macro
The macro is already tested very extensively. However, due to the nature
of procedural macros, most of these tests do not come in the form of
unit tests but integration tests, or don't exist in the same crate. As a
consequence, the macro shows up as poorly tested in code coverage.

That's unfair!

As a remedy, these unit tests.