perf: Faster `scale` by converting to XFE later
Keep the `BFieldElement` as a `BFieldElement` as long as possible. This
reduces the number of BFE multiplications when dealing when `Self =
Vec<XFieldElement>` in each loop-iteration from 15 to 6, as both
multiplications in the loop body are XFE * BFE. In the case of `Self =
Vec<BFieldElement>` this does not make a difference.