feat: serialize digest as hex for json, etc.
Implements custom serde Serialize and Deserializer so that Digest will
be hex encoded for human-readable formats like json, and will use
default serialization for all other formats (eg bincode).

This is useful because Digest will often be exposed as json in RPC
calls, and it is also used for *.dat files created by neptune-core.

It should be noted that hex encoding is 4 bytes smaller than the
default serialization for text formats because we omit 4 commas.

Also adds tests for serialization and improves existing tests