refactor(ZerofierTree): improve interface and mechanics
 - Add `new` function to `Leaf` and use it
 - Add `new` function to `Branch` and use it
 - Derive `PartialEq` for zerofier tree types and use equality operator to match enums
 - Remove clunky accessors
 - De-nest imports
 - Use `Box<_>` instead of `Arc<OnceLock<_>>`

The `Arc<OnceLock<_>>` serves two purposes
 - It introduces one level of indirection to create the recursive data
   structure.
 - It is thread-safe.

In fact, `Box<_>` can serve the same purposes.

The reason why you might want `Arc<OnceLock<_>>` sometimes is because
you don't want to set the data and you might not even know when it is
going to be set. In this case, we always set the data immediately.
It follows that we do not need interior mutability (even
once-mutability) and that the data structure is already thread-safe.

To be extra careful, this commit adds compile-time tests that the
`Sized + Send + Sync + Unpin` traits are implemented for the zerofier
tree data types.

Co-authored-by: Ferdinand Sauer <ferdinand@neptune.cash>