fix!: Fix platform-dependant digest encoding bug
Use a specified endianness, not the native as that depends on the system
you are executing the code on.

There are already tests that ensure that the bytes <-> digest
conversions are the identity operator when applied twice, but this test
was never run on a big-endian machine since we don't seem to have access
to one. Therefore, it is difficult to make a test that would have caught
this issue, so I haven't done that.

This closes #205.