refactor!(mmr): Remove leaf index from membership proof
With this commit, leaf index is removed from the struct
`MmrMembershipProof` which now only consists of authentication paths.
This is done for two reasons, one mathematical, one code and
performance-related:
1. A membership proof should not pertain to the data but only to the
   witness required to prove aspects of the data. And for an MMR which
   is an accumulator for a list, the index is considered data, not
   witness.
2. Downstream, in the mutator set implementation in neptune-core the inclusion
of leaf index in the mmr membership proof leads to duplicated data,
namely the leaf index being duplicated in, e.g., the `ChunkDictionary`.