refactor!(mmr): Add type for encapsulating leaf mutation
Add a `LeafMutation` type to encapsulate the action of mutating a leaf
in an MMR and thus to simplify the associated function signatures.

For performance reasons, the membership proof (authentication path)
associated with a leaf mutation does not need to be copied when
constructing a `LeafMutation` struct, as `LeafMutation` takes a
reference to a membership proof.