perf(`MmrSuccessorProof`): Drop MMR index arithmetic
Previously, `verify` would check that the MMR index of the landing
peak matches with that of the starting peak modulo incremental updates
with every step. MMR index arithmetic is expensive. This check would
happen in addition to checking that the height of the current node
matches that if the landing peak.

Now, `verify` checks that the peak index of the landing peak is not
larger than the peak index of the starting peak, in addition to
checking that the height of the landing peak matches that of the
current node. This does not lose any soundness because there is a
bijective map between (height, mmr index) and (height, peak index)
for peaks.