perf: Adjust cut-off values for fast/Lagrange interpolation
Alan and I ran some benchmarks on Mjolnir determining that Lagrange
interpolation is faster than "fast_interpolate" (sequential version) up
to a size of 2^{14}. The parallel version, "par_fast_interpolate" is
faster above a size of 2^{12}. So we adjust these thresholds to reflect
that.

We set the sequential threshold a bit lower though, also to 2^{12} bc
reasons. Close enough.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>