WIP: Improve `root_from_authentication_struct` algorithm
Leaf-indices are now absorbed in the sponge state to get the challenges.
Auth-struct indices are now read from non-determinism. They are bound by
the fact that they are required to be sorted, and that they must match
the parent indices in the last loop, where the `p` value is "restored"
by multiplying the inverses of the values from the first two loops.

Co-authored-by: Alan Szepieniec <alan@neptune.cash>