refactor!: Remove generic from `MerkleTree`
Previously, a Merkle tree could be built with any hash function
implementing the `AlgebraicHasher` trait. Only one such hash function
exists in the codebase: Tip5. The primary downstream consumers heavily
rely on Tip5 being used when building a Merkle tree. It is not expected
that any other hash function will replace Tip5. As a consequence, the
generality of `MerkleTree` is considered an unnecessary overhead.