refactor(polynomial)!: Generalize `evaluate()`
Previously, the method `Polynomial::evaluate()` only worked with
indeterminates and evaluation results of the same type as the
polynomial's coefficients.

Now, both the indeterminate and the evaluation result can be of any type
that is compatible with the coefficients and the operations performed
on them.

BREAKING CHANGE: Due to the new generality, additional type annotations
might be needed when using method `Polynomial::evaluate()`. If the old
behavior is preferred, method `Polynomial::evaluate_in_same_field()`
can be used as a drop-in replacement.