fix(BFieldCodec)!: Avoid possibly wrong `as` casts
If the target architecture's pointer width has fewer than 64 bits, `as`
casting a `u64` to a `usize` can be incorrect. It is better to present
an error than to continue the computation with incorrect data.

Also, mark `BFieldCodecError` as non-exhaustive to prevent changes like
these from being breaking in the future.

BREAKING CHANGE: Add variant `TryFromIntError` to `BFieldCodecError`.