fix!: Correctly parse “negative” `BFieldElement`s
For convenience and readability, some base field elements close to the
field-defining prime are represented through a negative number when
printed. For example, `BFieldElement::MAX` is printed as `-1` instead
of `18446744069414584320`.

Previously, such negative numbers were rejected when trying to parse
them as a `BFieldElement`, as those representations are not canonical.
Now, strings representing numbers in the open interval
`(-BFieldElement::P, BFieldElement::P)` are parsed correctly.