perf(MerkleTre): Parallelize over subtrees
Previously, building a Merkle tree in parallel meant that each layer was
built in parallel, with all worker threads synchronizing after each
layer. This is nice to implement but, as it turns out, has more overhead
than expected. A more performant strategy is to take a number of
subtrees of the big Merkle tree and compute all subtrees in parallel.

f