refactor!: Remove feature `mock`
The so-called “MockMMR” is a historic artifact that is unused in direct
dependencies. Turn it into a type that is accessible only from unit
tests and remove the feature `mock` to slim down the public API.

Replace `hashbrown::HashSet` with `std::collections::HashSet`, which
amounts to the same thing but allows removing the direct dependency on
`hashbrown`.

Other changes, only relevant for developers of twenty-first:
- rename the “MockMMR” to the more canonical “ArchivalMMR”.
- reflect the internals more accurately by renaming the collection of
  all nodes from “leafs” to “nodes”
- remove helper struct `MyVec<_>`
- replace some helper functions with sane initialize functions
- reduce visibility of some internal helper functions
- rework documentation in some places