test: tests can now be built
This adds the necessary Cargo.toml magic for `cargo test` to build tests
correctly, both #[test] and #[proptest].

However the wasm test harness ignores #[test] and #[proptest] so tests
are not actually run in that environment.

The source files will need to be modified to make the tests visible to
wasm test harness in a future commit.