chore: Turn `allow` lint levels into `expect`
This does not apply to code that gets generated by a macro, most
prominently, the `BFieldCodec` derive macro: there, we don't precisely
know the code that is being generated and _do_ want the behavior of
`allow`, not of `expect`.

Also, ease up on the `shadow_unrelated` lint. It's not nice to get a
compilation failure when quickly hacking around and accidentally
triggering this lint.

changelog: ignore