refactor!: Un-generalize some polynomial functions
In particular, we have observed that `Polynomial::scale` is only used
with an offset from the base field. The change cascades to both methods
`Polynomial::fast_coset_{evaluate, interpolate}`, which are only used
with a base field offset in downstream dependencies.

This change paves the way for some memoization in `Polynomial::scale`.